Refactor benchmakrs
